chore: move request into core
